Fixes issue where `gvp in CMD` breaks args
Running `gvp in CMD` breaks the arguments given to `CMD`. For example, if you try to pass a filename with spaces in it, the command will receive the filename broken up into separate arguments, when it should be just one (assuming you escape it correctly).

I'm not sure of the other ramifications of the change, but it fixes this issue.